% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_e_scale_symol_size.R
\name{e_scale_symbol_size}
\alias{e_scale_symbol_size}
\title{Scale the size of symbols in an echarts4r object}
\usage{
e_scale_symbol_size(e, width = 3, height = 3)
}
\arguments{
\item{e}{An echarts4r object}

\item{width}{The width of the symbols. Either a numeric value or the name of a column in the data.}

\item{height}{The height of the symbols. Either a numeric value or the name of a column in the data.}
}
\value{
An echarts4r object with the sizes of the symbols modified
}
\description{
Scale the size of symbols in an echarts4r object
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
iris |>
 dplyr::group_by(Species) |>
 dplyr::mutate(
   symbol_width = e_scale(Sepal.Width),
   symbol_height = e_scale(Sepal.Length)
 ) |>
 e_charts_("Sepal.Width") |>
 e_scatter_(
   "Sepal.Length",
   symbol = "rect"
 ) |>
 e_scale_symbol_size("symbol_width", "symbol_height") |>
 e_x_axis(min = 1.5) |>
 e_y_axis(min = 4)
\dontshow{\}) # examplesIf}
}
