% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api.R
\name{ecopi_api}
\alias{ecopi_api}
\title{Make API requests to the Ecopi API}
\usage{
ecopi_api(resource, ..., params = list(), new_data = list(), file_path)
}
\description{
This function sends an API request to the Ecopi API with the specified resource, parameters, and optional arguments.
}
